gestisciBirre/Sale dinamici

+++ src/main/java/it/uniroma3/siw/spring/model/Birra.java
@@ -22,7 +22,7 @@ public @Data class Birra {
 	private String birrificio;
 	private String colore;
 	private String descrizione;
-	private Float dimensione;
+	private Double dimensione;
 	private Integer prezzo;
 	
 	@ManyToMany
